Relocate unit and integration tests into __tests__ folder
